+++ Runtime/Core/Pooling/IPoolReleaser.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace NekoLib.Pooling
+{
+    internal interface IPoolReleaser
+    {
+        bool IsValid { get; }
+        void Despawn(Component instance);
+    }
+}
+
+