+++ Runtime/Core/Pooling/IPoolable.cs
@@ -6,3 +6,5 @@ namespace NekoLib.Pooling
         void OnDespawned();
     }
 }
+
+